+++ ydb/core/persqueue/pqtablet/cache/ut/ya.make
@@ -1,0 +1,7 @@
+UNITTEST_FOR(ydb/core/persqueue/pqtablet/cache)
+
+SRCS(
+    cache_eviction_ut.cpp
+)
+
+END()
