+++ ydb/core/persqueue/pqtablet/cache/ya.make
@@ -1,0 +1,20 @@
+LIBRARY()
+
+SRCS(
+    pq_l2_cache.cpp
+)
+
+
+
+PEERDIR(
+    ydb/core/base
+    ydb/core/keyvalue
+    ydb/library/logger
+)
+
+END()
+
+RECURSE_FOR_TESTS(
+    ut
+)
+
