+++ ydb/core/persqueue/pqtablet/partition/partition_monitoring.cpp
@@ -4,7 +4,7 @@
 #include "partition_compactification.h"
 #include "partition_util.h"
 #include "partition.h"
-#include <ydb/core/persqueue/read.h>
+#include <ydb/core/persqueue/pqtablet/cache/read.h>
 
 #include <ydb/core/base/appdata.h>
 #include <ydb/core/base/blobstorage.h>
