+++ ydb/core/persqueue/pqtablet/partition/partition_write.cpp
@@ -3,7 +3,7 @@
 #include <ydb/core/persqueue/pqtablet/common/logging.h>
 #include "partition_util.h"
 #include "partition.h"
-#include <ydb/core/persqueue/read.h>
+#include <ydb/core/persqueue/pqtablet/cache/read.h>
 
 #include <ydb/core/base/appdata.h>
 #include <ydb/core/base/blobstorage.h>
