+++ ydb/core/persqueue/pqtablet/ya.make
@@ -12,6 +12,7 @@ SRCS(
 
 PEERDIR(
     ydb/core/persqueue/pqtablet/common
+    ydb/core/persqueue/pqtablet/cache
     ydb/core/persqueue/pqtablet/partition
 )
